Fix slow linear by upgrading pytorch 2.4.1->2.6
We also use einsum to clean up the code, but einsum was still slow on
torch 2.4.1 for some reason. On the other hand, the for loop with simple
matrix multiplication worked fine on 2.4.1. Weird/Unexplained but do not
want to lose more time looking into this, given that switching to torch
2.6 fixes the issue.