Disable dynamo for circuit methods; stack marginals
We revert an older change where we stopped stacking marginals. This is
because we will no longer be doing full KL and the optimisation of not
stacking the marginals is not worth it given it breaks consistency of
the API - where tensors are always passed between functions instead of
lists of tensors.