Always mask for packing and allow KL
The existing code worked for packing because the examples include a
prompt which is padded with DO_NOT_PREDICT because we do not predict
prompts. However, if we were to predict prompts or load a pretraining
dataset, our code would fail because we would not be masking predictions
that take us across examples (i.e. condion on end of example 1 to
predict first tokens of example 2). This is now fixed - we always mask.